refactor: merge import meta glob dependency
