+++ crates/rspack_plugin_javascript/src/dependency/context/mod.rs
@@ -50,7 +50,6 @@ mod amd_require_context_dependency;
 mod common_js_require_context_dependency;
 mod import_context_dependency;
 mod import_meta_context_dependency;
-mod import_meta_glob_dependency;
 mod import_meta_resolve_context_dependency;
 mod require_context_dependency;
 mod require_resolve_context_dependency;
@@ -66,7 +65,6 @@ pub use import_context_dependency::{ImportContextDependency, ImportContextDepend
 pub use import_meta_context_dependency::{
   ImportMetaContextDependency, ImportMetaContextDependencyTemplate,
 };
-pub use import_meta_glob_dependency::{ImportMetaGlobDependency, ImportMetaGlobDependencyTemplate};
 pub use import_meta_resolve_context_dependency::{
   ImportMetaResolveContextDependency, ImportMetaResolveContextDependencyTemplate,
 };
