+++ crates/rspack_plugin_javascript/src/parser_plugin/mod.rs
@@ -12,7 +12,6 @@ mod esm_import_dependency_parser_plugin;
 mod esm_top_level_this_plugin;
 mod exports_info_api_plugin;
 mod import_meta_context_dependency_parser_plugin;
-mod import_meta_glob_dependency_parser_plugin;
 mod import_meta_plugin;
 mod import_parser_plugin;
 mod initialize_evaluating;
@@ -55,7 +54,6 @@ pub(crate) use self::{
   esm_top_level_this_plugin::ESMTopLevelThisParserPlugin,
   exports_info_api_plugin::ExportsInfoApiPlugin,
   import_meta_context_dependency_parser_plugin::ImportMetaContextDependencyParserPlugin,
-  import_meta_glob_dependency_parser_plugin::ImportMetaGlobDependencyParserPlugin,
   import_meta_plugin::{ImportMetaDisabledPlugin, ImportMetaPlugin},
   import_parser_plugin::{ImportParserPlugin, ImportsReferencesState},
   initialize_evaluating::InitializeEvaluating,
